libfuse: Add posix ACL support
Add support for posix ACLs by intercepting posix ACL xattrs in
getxattr/setxattr/removexattr and redirecting these to
getacl/setacl operations. The filesystems are responsible for
fetching the ACLs and converting to/from whatever representation
is used on disk and also for updating the file mode as needed.
Helper functions are supplied for converting to and from the fuse
ACL structure to the posix layout and for calculating the
equivalent file mode from the ACL.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>