lib/{mount.c,mount_util.c}: Depend on IGNORE_MTAB instead of __NetBSD__.
Define IGNORE_MTAB in mount.c and mount_util.c if __NetBSD__ is defined.

In mount.c and mount_util.c questions when (if we build on NetBSD) and
why (if we ignore mtab) were mixed.  Depending solely on IGNORE_MTAB in
the code part is better, because it makes the intent (and the code)
clearer.

The #else branch of #ifdef __NetBSD__ in mount_util.c, that undefines
IGNORE_MTAB if it was already defined, has been introduced to preserve
status quo, which also exposes the issue that --disable-mtab was not
truly respected and mtab_needs_update() was called on non-NetBSD
platforms.  (It will be fixed in the next commit.)

Machine code generated from both files remains the same as before.

util/fusermount.c's preprocessor directives also check IGNORE_MTAB, but
__NetBSD__ is nowhere inspected there, and because I don't have NetBSD
and I don't know NetBSD mounting behavior, I refrained from touching
this file and making -D__NetBSD__ -> -DIGNORE_MTAB a global thing.