Allow inode cache invalidation in high-level API
We introduce a new routine to the high-level FUSE API,
fuse_notify_inval_fs_inode, which is responsible for invalidating the
cached data for an inode.

The routine uses a new mapping of the "FS inodes" (inode numbers
returned in the 'st_ino' field from 'getattr') into the nodes in the
high-level FUSE API. Naturally, because multiple paths may refer to
the same inode on the file system due to hard links, the mapping is
one-to-many, and is implemented by means of a hash table. Internally,
the high-level API routine calls the low-level routine
fuse_notify_inval_inode.

This is useful for network-based file systems which use the high-level
FUSE API. Up to now, there was no way for those file systems to notify
the kernel about external changes.