Stop using MESON_SOURCE_ROOT in install script
When libfuse is used as a meson subproject, MESON_SOURCE_ROOT is set to
the parent package source directory in install script environment, thus
breaking source file references. There doesn't seem to be a preferred
way of accessing the current project's source directory from such a script
at the moment [1]. This patch works around this by passing the current
project source directory as an argument of the script.

[1] https://github.com/mesonbuild/meson/issues/4399