fuse-lowlevel: set pipe size to max
on failure to set the pipe size, set it to the maximum allowed by the
kernel.

If the first request required more than the maximum allowed, the
can_grow flag would be reset thus preventing any further resize.

Grow the pipe to the maximum allowed to increase the likelihood of
using splice for successive requests instead of falling back to
read/write.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>