fuse_types.h: rename standard types
This commit renames standard system types to equivalent FUSE types.
For example, `struct stat` is renamed to `struct fuse_stat`.

This allows libfuse to be potentially ported to platforms which either
do not have these types or that have antiquated versions of these types
(e.g. Windows). On the platforms which are the current primary targets
of libfuse (Linux, FreeBSD, etc.) the new names are simple macros over
the standard types:

    #define fuse_stat stat