Implement Atomic lookup + open
There are certain places where we can avoid costly lookup calls
into libfuse from fuse kernel.
1) When we go for opening the file with O_CREAT flags.
2) When we are opening an already existing file.

In first case, since we are passing O_CREAT, its very much likely
that file does not exist yet so lookup performed before creating a
file can be avoided. Instead this lookup can be performed as part of
open call made into FUSE user space via libfuse.

In second case, we avoid lookup which is done before opening the file
/directory, instead we revalidate the inode (non-directories) in fuse
kernel once fresh attributes are received as part of this atomic open.

Now with these changes, user space when receiving atomic open is
supposed to perform lookup on the file/dir, fill in the attributes
and open the file. These filled in attributes are used by fuse
kernel to make indoe stand or revalidate the already existing inode.

Note: In second case we avoid lookups only for non-dir and when
FUSE is not mounted with default permissions otherwise lookup
can't be avoided as we need to fetch fresh attributes of the
inode for permission checks.