Implement lookups during file create
There are certain places where we can avoid costly lookup calls
into libfuse from fuse kernel. One of such place is When we go
for opening the file with O_CREAT flags.

Since we are going to create a file (passing O_CREAT in open),
its very much likely that file does not exist yet so lookup
performed before creating a file can be avoided. Instead this
lookup can be performed as part of create call itself.

Now with these changes, user space when receiving extended create is
supposed to create the file if it does not exist/perform lookup on
the file if already existing, fill in the attributes, set 'file_created'
bit in 'struct fuse_file_info' and open the file. These filled in
attributes are used by fuse kernel to make inode stand/revalidate.