passthrough_ll: Use a btree to store lo_inode
passthrough_ll has an issue with benchmarks and also with
xfs_test generic/531 - it uses a double linked list to
store the file handle (inode) with an O(N) search time.
Using it for benchmarks with many files already caused a bit
confusion and we had switched to passthrogh_hp for that reason.
However, passthrough_hp is not suitiable to run xfstests, as
it does not accept standard mount options and fails to start up
(as it uses its own option parser instead of fuse_parse_cmdline()).

This is an attempt to make passthrough_ll faster by using a btree
to store the inode. Btree code is taken from the OpenBSD project.