xfstest examples: Remove SCRATCH_SOURCE and TEST_SOURCE
There are probably not many people running xfstests yet
and then these are just example scripts - we can probably
safely remove these variables and simplidy the scripts
with that.

This also adds a check if the mount helper got the passthrough
source directory and aborts with an error if not.