ci-build.sh: Always install and add s-bit for fusermount3
As per pull #898, fusermount3 had a severe issue that
should have been detected by ASAN. I guess tests used
the system default and not the sanitized binary.
Order of execution of fusermount3 is to try
1) full install path
if that fails
2) just fusermount3

So tests should be fixed by installing libfuse, setting the s-bit
on fusermount3 and then to run the tests.