Add "FUSE_NO_MTAB_UPDATE" env to disable mtab updates
This adds the env FUSE_NO_MTAB_UPDATE variable to be able
to disable mtab updates without recompiling libfuse.
The value needs to be set to 1.

This is an addition to commit 74b1df2e, which removed
a check for mtab updates, if mtab is a symlink.
Although not related to the symlink, this caused a regression
in our code, as the symlink update was done through fork.
The env variable is a run time workaround for such cases and
to avoid the requirement to disable mtab updates at libfuse
configuration/compilation time.