Add the fuse version as string and integers to the config file
This will be useful to store into binaries which libfuse version a
file system was compiled with.

Defining the version in fuse_common.h, is removed, it is defined
through meson and provided by "libfuse_config.h".
The part that might be revised later is how the version is defined
in meson - ideally it would set integers and create the version
string from these integers. However, meson requires that "project"
is the first meson.build keyword - with that it requires to
set the version from a string and then major/minor/hotfix integers
are created from string split.

This introduces the 317 ABI, with fuse_session_new_317. And
fuse_new_317

XXX: Do not merge, fuse_new_317 not implemented yet.