Use a single bit for "no_interrupt"
This is an addition to commit cef8c8b24
"Add support for no_interrupt"
and uses a single bit - the variable is used as boolean.

This should allow us to add more booleans into the "unsigned" in
the future. Though I'm not sure yet how to handle "$arch" that does
not have unsigned as 32bit value.