Add request timeout mount option
There is a patch upstream for supporting kernel-enforced timeouts
for fuse requests. This sets an upper bound on how long a user
can expect a call to take.

This commit adds this option in libfuse and adds a test,
test_timeout.c, which can be used for exercising the paths where
a request times out.

Signed-off-by: Joanne Koong <joannelkoong@gmail.com>