Update struct fuse_config (int32_t/uint8_t/paddings)
Convert
unsigned -> uint32_t
int      -> int32_t

Also convert several variables from int/uint to uint8_t, as
they are used as boolean type. Ideal would be to use single bits,
but some of these parameters are exposed as config options in
fuse.c 'fuse_lib_opts' and accessed with the help of offsetof(),
bit fields are not working with that. For consistency all
these variables are kept as uint8_t, even those that are not
exposed. The C 'bool' type is not used, as the exact width is not
defined by the C standard.

Also reserve bytes ('padding') at the end of the struct, for
future ABI compatibility.

The struct was optimized with pahole to reserve all holes with
paddings for future usage.

The fields 'fmask' and 'dmask' (new in 3.17) have been moved
closer to 'set_mode' and 'umask' as they are related to these.

The fields 'ac_attr_timeout_set' and 'ac_attr_timeout' have
been moved out of the range of uint8 (bool like) fields.